overall clean up
