+++ FloorplanToBlenderLib/transform.py
@@ -1,5 +1,6 @@
 import math
 import cv2
+import numpy as np
 '''
 Transform
 This file contains functions for transforming data between different formats.
